pyproject.toml
[project]
name = "seven"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[tool.pytest.ini_options]
pythonpath = ["."]
